src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataclasses>=0.8",
    "datetime>=5.5",
    "mysql-connector>=2.2.9",
    "mysql-connector-python>=9.3.0",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "pyqt5>=5.15.11",
    "typing>=3.10.0.0",
]
